Weight services

+++ livestock-tracker.abstractions/Services/Medical/IMedicalTransactionSearchService.cs
@@ -5,7 +5,7 @@ namespace LivestockTracker.Abstractions.Services.Medical
     /// <summary>
     /// / Provides fetch and pagination services for medical transactions.
     /// </summary>
-    public interface IMedicalTransactionSearchService : IPagedFetchAsyncService<IMedicalTransaction>, IFetchAsyncService<IMedicalTransaction, int>
+    public interface IMedicalTransactionSearchService : IPagedFetchAsyncService<IMedicalTransaction>, IFetchAsyncService<IMedicalTransaction, long>
     {
     }
 }
